requirements.txt
nose
PyYAML
boto3
